# Conflicts:
#	x-pack/plugins/osquery/server/plugin.ts

+++ x-pack/plugins/fleet/server/integration_tests/docker_registry_helper.ts
@@ -8,9 +8,12 @@
 import { spawn } from 'child_process';
 import type { ChildProcess } from 'child_process';
 
+import pRetry from 'p-retry';
 import fetch from 'node-fetch';
 
 const delay = (ms: number) => new Promise((resolve) => setTimeout(resolve, ms));
+
+const DOCKER_START_TIMEOUT = 5 * 60 * 1000; // 5 minutes
 
 export function useDockerRegistry() {
   const packageRegistryPort = process.env.FLEET_PACKAGE_REGISTRY_PORT || '8081';
@@ -21,7 +24,7 @@ export function useDockerRegistry() {
 
   let dockerProcess: ChildProcess | undefined;
   async function startDockerRegistryServer() {
-    const dockerImage = `docker.elastic.co/package-registry/distribution@sha256:de952debe048d903fc73e8a4472bb48bb95028d440cba852f21b863d47020c61`;
+    const dockerImage = `docker.elastic.co/package-registry/distribution@sha256:8b4ce36ecdf86e6cfdf781d9df8d564a014add9afc9aec21cf2c5a68ff82d3ab`;
 
     const args = ['run', '--rm', '-p', `${packageRegistryPort}:8080`, dockerImage];
 
@@ -32,8 +35,9 @@ export function useDockerRegistry() {
       isExited = true;
     });
 
-    let retries = 0;
-    while (!isExited && retries++ <= 20) {
+    const startedAt = Date.now();
+
+    while (!isExited && Date.now() - startedAt <= DOCKER_START_TIMEOUT) {
       try {
         const res = await fetch(`http://localhost:${packageRegistryPort}/`);
         if (res.status === 200) {
@@ -46,8 +50,12 @@ export function useDockerRegistry() {
       await delay(3000);
     }
 
+    if (isExited && dockerProcess.exitCode !== 0) {
+      throw new Error(`Unable to setup docker registry exit code ${dockerProcess.exitCode}`);
+    }
+
     dockerProcess.kill();
-    throw new Error('Unable to setup docker registry');
+    throw new pRetry.AbortError('Unable to setup docker registry after timeout');
   }
 
   async function cleanupDockerRegistryServer() {
@@ -57,8 +65,11 @@ export function useDockerRegistry() {
   }
 
   beforeAll(async () => {
-    jest.setTimeout(5 * 60 * 1000); // 5 minutes timeout
-    await startDockerRegistryServer();
+    const testTimeout = 5 * 60 * 1000; // 5 minutes timeout
+    jest.setTimeout(testTimeout);
+    await pRetry(() => startDockerRegistryServer(), {
+      retries: 3,
+    });
   });
 
   afterAll(async () => {
